Update GCD & LCM.cpp

+++ Support/GCD & LCM.cpp
@@ -1,7 +1,7 @@
-long long gcd(long long a, long long b) {
+ll gcd(ll a, ll b) {
     if(b != 0) return gcd(b, a%b);
     else return a;
 }
-long long lcm(long long a, long long b) {
+ll lcm(ll a, ll b) {
     return a*b / gcd(a,b);
 }
